Updated for PRO Version Spring20
Just updated the Encryption Method.
